moved components over
